requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
aiohttp==3.10.5
pydantic[email]==2.9.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-multipart==0.0.12
